vehicle-sim/requirements.txt
tcconfig~=0.28.0
fastapi~=0.95.1
uvicorn~=0.22.0
Faker~=17.6.0
paho-mqtt~=1.6.1
pika~=1.3.2
